requirements.txt
flask
gunicorn
openai
pymongo
python-dotenv
